+++ lib/noticed.rb
@@ -12,15 +12,17 @@ module Noticed
   autoload :NotificationChannel, "noticed/notification_channel"
 
   module DeliveryMethods
-    autoload :Base, "noticed/delivery_methods/base"
     autoload :ActionCable, "noticed/delivery_methods/action_cable"
+    autoload :Base, "noticed/delivery_methods/base"
     autoload :Database, "noticed/delivery_methods/database"
     autoload :Email, "noticed/delivery_methods/email"
-    autoload :Slack, "noticed/delivery_methods/slack"
+    autoload :Ios, "noticed/delivery_methods/ios"
     autoload :MicrosoftTeams, "noticed/delivery_methods/microsoft_teams"
+    autoload :Slack, "noticed/delivery_methods/slack"
     autoload :Test, "noticed/delivery_methods/test"
     autoload :Twilio, "noticed/delivery_methods/twilio"
     autoload :Vonage, "noticed/delivery_methods/vonage"
+    autoload :Fcm, "noticed/delivery_methods/fcm"
   end
 
   def self.notify(recipients:, notification:)